unify R.clone, R.cloneObj, and R.cloneDeep
